backend/requirements.txt
flask
flask-cors
supabase
python-dotenv
flasgger
geoip2
moviepy
openai-whisper
langchain
langchain-google-genai
transformers
adapters
torch
datasets
